fixed the table schema

+++ cbt-app/database/migrations/2025_05_07_115405_create_students_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('other_name')->nullable();
             $table->string('email')->unique();
             $table->string('password');  
-            $table->numeric('phone')->nullable();
+            $table->integer('phone')->nullable();
             $table->enum('gender', ['male', 'female']);
             $table->string('state')->nullable();
             $table->string('country')->nullable();
